% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{List Available r4subdata Datasets}
\usage{
list_datasets()
}
\value{
A tibble with columns: \code{name}, \code{description}, \code{n_rows}, \code{n_cols}.
}
\description{
Returns a summary of all datasets included in the r4subdata package.
}
\examples{
list_datasets()

}
