% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4subdata-package.R
\docType{package}
\name{r4subdata-package}
\alias{r4subdata}
\alias{r4subdata-package}
\title{r4subdata: Example Datasets for Clinical Submission Readiness}
\description{
Provides realistic synthetic example datasets for the R4SUB (R for Regulatory Submission) ecosystem. Includes a pharma study evidence table, ADaM (Analysis Data Model) and SDTM (Study Data Tabulation Model) metadata following CDISC (Clinical Data Interchange Standards Consortium) conventions (\url{https://www.cdisc.org}), traceability mappings, a risk register based on ICH (International Council for Harmonisation) Q9 quality risk management principles (\url{https://www.ich.org/page/quality-guidelines}), and regulatory indicator definitions. Designed for demos, vignettes, and package testing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/R4SUB/r4subdata}
  \item Report bugs at \url{https://github.com/R4SUB/r4subdata/issues}
}

}
\author{
\strong{Maintainer}: Pawan Rama Mali \email{prm@outlook.in} [copyright holder]

}
\keyword{internal}
